requirements.txt
#pip3 install -r requirements.txt

Appium-Python-Client==4.5.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile ==2.0.0
execnet==2.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pip==25.0.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-django==4.10.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata== 2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto== 1.2.0